requirements.txt
python-telegram-bot[job-queue]>=20
python-dotenv>=1.0
